requirements.txt
torch==1.9.0
torchvision==0.10.0
numpy
pillow
tensorboardX
opencv-python
plyfile